+++ changelog.md
@@ -10,6 +10,7 @@ A change like this was requested for a while...
 - `hata.ext` directory now handles sub-file imports better.
 - Add `weak` parameter to `bind`.
 - Add `weak_cache_size` parameter to `bind`.
+- Add `Guild.vanity_url`.
 
 ##### ext.slash